requirements.txt
# ============================================================
# Requirements — Desenvolvimento Local (fora do Docker)
# ============================================================

# Data Engineering
apache-airflow==2.9.3
dbt-core==1.8.7
dbt-duckdb==1.8.3
duckdb==1.1.3

# Data Science
scikit-learn==1.5.2
xgboost==2.1.3
lightgbm==4.5.0
shap==0.46.0
numpy==1.26.4
pandas==2.2.3
pyarrow==18.1.0
scipy==1.14.1

# MLflow
mlflow==2.14.3
boto3==1.35.86  # para MinIO/S3

# FastAPI
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3

# LLM
anthropic==0.40.0

# Visualização / Notebooks
matplotlib==3.9.3
seaborn==0.13.2
plotly==5.24.1
ipykernel==6.29.5

# Testes
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1  # TestClient async

# Utilitários
python-dotenv==1.0.1
loguru==0.7.2
great-expectations==0.18.15  # validação de dados (opcional)
